+++ packages/data-models/index.ts
@@ -1,4 +1,5 @@
 export * from "./appConfig";
+export * from "./assets.model";
 export * from "./db.model";
 export * from "./deployment.model";
 export * from "./fields";
